refactor(docs): pagination dx
