+++ apps/docs/content/components/pagination/boundaries.ts
@@ -1,28 +1,4 @@
-const App = `import {Pagination, Button} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex flex-col gap-5">
-      <p>1 Boundary (default)</p>
-      <Pagination
-        total={10}
-        color="secondary"
-      />
-      <p>2 Boundaries</p>
-      <Pagination
-        total={10}
-        boundaries={2}
-        color="secondary"
-      />
-      <p>3 Boundaries</p>
-      <Pagination
-        total={10}
-        boundaries={3}
-        color="secondary"
-      />
-    </div>
-  );
-}`;
+import App from "./boundaries.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
