+++ apps/docs/content/components/pagination/colors.ts
@@ -1,17 +1,4 @@
-const App = `import {Pagination} from "@nextui-org/react";
-
-export default function App() {
-
-  const colors = ["primary", "secondary", "success", "warning", "danger"]
-
-  return (
-    <div className="flex flex-wrap gap-4 items-center">
-      {colors.map((color) => (
-        <Pagination key={color} total={10} initialPage={1} color={color} />
-      ))}
-    </div>
-  );
-}`;
+import App from "./colors.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
