+++ apps/docs/content/components/pagination/controlled.raw.jsx
@@ -1,0 +1,30 @@
+import {Pagination, Button} from "@nextui-org/react";
+
+export default function App() {
+  const [currentPage, setCurrentPage] = React.useState(1);
+
+  return (
+    <div className="flex flex-col gap-5">
+      <p className="text-small text-default-500">Selected Page: {currentPage}</p>
+      <Pagination color="secondary" page={currentPage} total={10} onChange={setCurrentPage} />
+      <div className="flex gap-2">
+        <Button
+          color="secondary"
+          size="sm"
+          variant="flat"
+          onPress={() => setCurrentPage((prev) => (prev > 1 ? prev - 1 : prev))}
+        >
+          Previous
+        </Button>
+        <Button
+          color="secondary"
+          size="sm"
+          variant="flat"
+          onPress={() => setCurrentPage((prev) => (prev < 10 ? prev + 1 : prev))}
+        >
+          Next
+        </Button>
+      </div>
+    </div>
+  );
+}
