+++ apps/docs/content/components/pagination/custom-styles.raw.jsx
@@ -1,0 +1,15 @@
+import {Pagination} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Pagination
+      classNames={{
+        wrapper: "gap-0 overflow-visible h-8 rounded border border-divider",
+        item: "w-8 h-8 text-small rounded-none bg-transparent",
+        cursor:
+          "bg-gradient-to-b shadow-lg from-default-500 to-default-800 dark:from-default-300 dark:to-default-100 text-white font-bold",
+      }}
+      total={10}
+    />
+  );
+}
