+++ apps/docs/content/components/pagination/disabled.raw.jsx
@@ -1,0 +1,5 @@
+import {Pagination} from "@nextui-org/react";
+
+export default function App() {
+  return <Pagination isDisabled initialPage={1} total={10} />;
+}
