+++ apps/docs/content/components/pagination/initial-page.raw.jsx
@@ -1,0 +1,5 @@
+import {Pagination} from "@nextui-org/react";
+
+export default function App() {
+  return <Pagination color="warning" initialPage={3} total={10} />;
+}
