+++ apps/docs/content/components/pagination/loop.raw.jsx
@@ -1,0 +1,5 @@
+import {Pagination} from "@nextui-org/react";
+
+export default function App() {
+  return <Pagination loop showControls color="success" initialPage={1} total={5} />;
+}
