+++ apps/docs/content/components/pagination/loop.ts
@@ -1,10 +1,4 @@
-const App = `import {Pagination} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Pagination loop showControls color="success" total={5} initialPage={1} />
-  );
-}`;
+import App from "./loop.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
