+++ apps/docs/content/components/pagination/shadow.ts
@@ -1,10 +1,4 @@
-const App = `import {Pagination} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Pagination showShadow color="warning" total={10} initialPage={1} />
-  );
-}`;
+import App from "./shadow.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
