+++ apps/docs/content/components/pagination/sizes.ts
@@ -1,17 +1,4 @@
-const App = `import {Pagination} from "@nextui-org/react";
-
-export default function App() {
-
-  const sizes = ["sm", "md", "lg"]
-
-  return (
-    <div className="flex flex-wrap gap-4 items-center">
-      {sizes.map((size) => (
-        <Pagination key={size} total={10} initialPage={1} size={size} />
-      ))}
-    </div>
-  );
-}`;
+import App from "./sizes.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
