+++ apps/docs/content/components/pagination/variants.ts
@@ -1,17 +1,4 @@
-const App = `import {Pagination} from "@nextui-org/react";
-
-export default function App() {
-
-  const variants = ["flat", "bordered", "faded", "light"]
-
-  return (
-    <div className="flex flex-wrap gap-4 items-center">
-      {variants.map((variant) => (
-        <Pagination key={variant} total={10} initialPage={1} variant={variant} />
-      ))}
-    </div>
-  );
-}`;
+import App from "./variants.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
